refactor(cli): extract glyph helpers, dedup _load_wallet, drop dead code
Three behaviour-preserving cleanups to the glyph CLI module:

* Move the glyph-specific private helpers out of glyph_cmds.py into a new
  glyph_helpers.py — metadata parsing/scaffolding (_read_metadata_file,
  _TEMPLATE_TYPES, _scaffold_for), the pre-broadcast confirmation summary
  (_BroadcastSummary, _confirm_or_abort, _metadata_summary), the reveal
  unlock-script builder (_build_glyph_unlock), and ref parsing (_parse_ref,
  _try_extract_ft_ref). glyph_cmds.py imports them back; the file drops from
  1214 to ~930 lines and now reads as command flow.

* Dedup _load_wallet. It was duplicated near-verbatim in query_cmds.py and
  glyph_cmds.py; the single canonical copy now lives in prompts.py (which
  already owns the mnemonic/passphrase prompts it orchestrates) and both
  command modules import it. The body is unchanged — only the docstring.

* Remove _select_funding_utxo — dead code with zero references repo-wide.

All moves are verbatim (function bodies byte- and AST-identical to before).
ruff + format clean, no import cycle, full suite green (3734 passed).
